api-collector/api/users/dependencies.py
from fastapi import Depends, Request
from jose import ExpiredSignatureError, JWTError, jwt

from api.config import settings
from api.exceptions import (
    IncorrectTokenFormatException,
    TokenAbsentException,
    TokenExpiredException,
    UserIsNotPresentException,
)
from api.users.dao import UserDAO


def get_token(request: Request):
    token = request.cookies.get("asmi_access_token")
    if not token:
        raise TokenAbsentException
    return token


async def get_current_user(token: str = Depends(get_token)):
    try:
        payload = jwt.decode(
            token, settings.SECRET_KEY, settings.ALGORITHM
        )
    except ExpiredSignatureError:
        # Как позже выяснилось, ключ exp автоматически проверяется
        # командой jwt.decode, поэтому отдельно проверять это не нужно
        raise TokenExpiredException
    except JWTError:
        raise IncorrectTokenFormatException
    user_id: str = payload.get("sub")
    if not user_id:
        raise UserIsNotPresentException
    user = await UserDAO.get_one_or_none(id=int(user_id))
    if not user:
        raise UserIsNotPresentException

    return user
